Commanditaires & Opérateurs ASPE par date Descriptif opération

Interroge la base ASPE via un fichier YAML de connexion et extrait, pour une station ou un point ASPE, les acteurs :

commanditaire_libelle_aspe (paneau du haut),

operateur_libelle_aspe (paneau du bas). Puis les modes de pêche

protocole_peche (paneau du haut),

moyen_prospection (paneau du bas),

graph_descriptif_operation(
  yaml_config,
  code_station = NULL,
  code_point_prelevement_aspe = NULL,
  annee_debut = 1950,
  annee_fin = as.integer(format(Sys.Date(), "%Y")),
  n_last = 10,
  titre = "Descriptif opération",
  normalize_accents = TRUE
)

Arguments

titre

Titre du graphique (défaut : "Descriptif opération").

Value

Un objet ggplot2 facetté (2 lignes : Commanditaire en haut, Opérateur en bas), avec en y les acteurs et, pour chaque date en x, un point si l'acteur est présent.

Details

L'axe des x représente les dates (facteur ordonné), l'axe des y les acteurs (une ligne par valeur). Les points indiquent la présence de l'acteur pour la date donnée. Les couleurs différencient le type d'acteur (Commanditaire vs Opérateur) et le graphique est facetté en quatre panneaux empilés partie haute : (haut = Commanditaire ; bas = Opérateur). partie basse (haut = protocole_peche, bas = moyen_prospection)

Examples

if (FALSE) { # \dontrun{
graph_descriptif_operation(
  yaml_config = "config/connexion.yml",
  code_station = "04216000",
  annee_debut = 2000,
  n_last = 12
)
} # }