Histogramme des surfaces de pêche ASPE pour une station

Cette fonction interroge la base ASPE via un fichier YAML de connexion et extrait les profondeurs de prélèvement (surface_calculee) issues de la table qe.aspe_operations pour une station donnée (via code_station) ou pour un point ASPE (code_point_prelevement_aspe).

plot_surface_pechee_histogram(
  yaml_path,
  code_station = NULL,
  code_point_prelevement_aspe = NULL,
  annee_debut = 1950,
  annee_fin = as.integer(format(Sys.Date(), "%Y")),
  n_last = 10,
  titre = "Surface pêchée en m²"
)

Arguments

yaml_path

Chemin vers le fichier YAML contenant les informations de connexion PostgreSQL.

code_station

[character(1)]\ Code Sandre de la station. Si renseigné conjointement avec code_point_prelevement_aspe, il est prioritaire (avec warning).

code_point_prelevement_aspe

[character(1)]\ Code du point de prélèvement ASPE utilisé si code_station n'est pas fourni.

annee_debut

[numeric(1)]\ Année minimale des opérations à prendre en compte (défaut : 1950).

annee_fin

[numeric(1)]\ Année maximale (défaut : année courante).

n_last

[integer(1)]\ Nombre de dernières dates d’opérations (distinctes) à représenter.

titre

[character(1)]\ Titre du graphique (défaut : "Prof. moy. station (m)").

Value

Un objet ggplot2 représentant un histogramme des profondeurs de prélèvement. Retourne NULL si aucune donnée n’est disponible.

Details

Si les deux identifiants sont fournis, code_station est prioritaire et un warning est renvoyé.

Les données sont filtrées entre annee_debut et annee_fin, puis seules les n_last dernières dates d’opération distinctes sont conservées.

La fonction produit un graphique ggplot2 de type histogramme (barres verticales) où la hauteur des barres représente la profondeur (en mètres). Les dates sont affichées au format \ et pivoteées de 90° pour faciliter la lecture.

La table interrogée est : qe.aspe_operations, champ profondeur.

Les dates d’opération (date_operation) sont converties en classe Date, puis ordonnées avant d’être transformées en facteur pour un affichage strictement chronologique dans le graphique.

Examples

if (FALSE) { # \dontrun{
plot_surface_pechee_histogram(
 yaml_path = "C:/workspace/gwilenalim/yaml/config.yml",
 code_station = "04216050",
 annee_debut  = 2000,
 annee_fin    = 2025
)
} # }